adds parameter and typedef options and parameter removal

+++ help/csharp/HowToHandshakeUtils/RCPClient.cs
@@ -143,7 +143,7 @@ namespace RCP
 
                 case RcpTypes.PacketTypes.Remove:
                         {
-                            var id = (short)packet.Data;
+                            var id = (int)packet.Data;
 
                             if (FParams.ContainsKey(id))
                                 RemoveParameter(FParams[id]);
